pyproject.toml
[project]
name = "fastapi-boilerplate"
version = "0.1.0"
description = "A modern FastAPI microservices boilerplate"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "pydantic==2.10.0",
    "pydantic-settings==2.6.0",
    "email-validator==2.2.0",
    "SQLAlchemy==2.0.36",
    "alembic==1.14.0",
    "psycopg2-binary==2.9.10",
    "mysqlclient==2.2.4",
    "python-jose[cryptography]==3.4.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==5.0.0",
    "cryptography==46.0.3",
    "httpx==0.28.1",
    "redis[hiredis]==5.2.0",
    "celery==5.6.0",
    "APScheduler==3.11.0",
    "python-socketio==5.14.0",
    "websockets==14.1",
    "fs==2.4.16",
    "fs-s3fs==1.1.1",
    "boto3==1.35.29",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.18",
    "python-dateutil==2.9.0.post0",
    "click==8.1.7",
]

[dependency-groups]
dev = [
    "pytest==8.3.4",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-python-dateutil",
    "types-redis",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
exclude = ["venv", ".venv", "alembic"]

[tool.pytest.ini_options]
testpaths = ["app/tests"]
python_files = ["test_*.py"]
addopts = "--cov=app --cov-report=term-missing"
